Share code between NettyHandlingSpec and ServerBenchmark
* Refactor code in NettyHandlingSpec into NettyRunners and
  NettySpecification.
* Move ServerBenchmark into Play-Integration-Tests so that code can
  be reused.
* Rewrite ServerBenchmark to use logic from new NettyRunners class.
  The benchmark is now slightly slower than before, probably because
  a separate channel is created for each request. This was a
  necessary change as part of the refactor, but it also gives a
  more realistic test.
