+++ framework/project/Dependencies.scala
@@ -116,7 +116,6 @@ object Dependencies {
 
     "org.scala-lang" % "scala-reflect" % BuildSettings.buildScalaVersion,
 
-    "org.databene" % "contiperf" % "2.2.0" % "test",
     "junit" % "junit" % "4.11" % "test")
 
   val link = Seq(
@@ -196,6 +195,10 @@ object Dependencies {
     ("org.fluentlenium" % "fluentlenium-festassert" % "0.8.0")
       .exclude("org.jboss.netty", "netty")
       .exclude("com.google.guava","guava"))
+
+  val integrationTestDependencies = Seq(
+    "org.databene" % "contiperf" % "2.2.0" % "test"
+  )
 
   val playCacheDeps = Seq(
     "net.sf.ehcache" % "ehcache-core" % "2.6.6",
